Add suppressor scaffolding
